# doc/backend-OpenAPIdoc.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Data Filter Web Application",
        "description": "A travel time data filter application for the City of Toronto. Made by University of Toronto CSC301 Project Team 22."
    },
    "servers": [
        {
            "url": "http://127.0.0.1"
        },
        {
            "url": "http://backendtest-env.eba-aje3qmym.ca-central-1.elasticbeanstalk.com/"
        },
        {
            "url": "http://backendproduction-env.eba-5qbcpngm.ca-central-1.elasticbeanstalk.com/"
        }
    ],
    "paths": {
        "/": {
            "get": {
                "summary": "Project Title",
                "operationId": "getProjectTitle",
                "responses": {
                    "200": {
                        "description": "This is project title.",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/closest-node/{longitude}/{latitude}": {
            "parameters": [
                {
                    "name": "longitude",
                    "in": "path",
                    "required": true,
                    "description": "The longitude of the given coordinate.",
                    "schema": {
                        "type": "string"
                    }
                },
                {
                    "name": "latitude",
                    "in": "path",
                    "required": true,
                    "description": "The latitude of the given coordinate.",
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "get": {
                "summary": "Get the 10 closest nodes from the given coordinate",
                "operationId": "getClosestNode",
                "responses": {
                    "200": {
                        "description": "Ten closest nodes from given coordinate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Node"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/link-nodes": {
            "post": {
                "summary": "Get the shortest link between the given series of nodes.",
                "operationId": "getLinkBetweenNodes",
                "requestBody": {
                    "description": "The series of nodes",
                    "required": true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/LinkNodesRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The shortest series of links between the series of nodes.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {"$ref": "#/components/schemas/SingleLinkPath"},
                                        {"$ref": "#/components/schemas/MultiLinkPath"}
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/travel-data-file": {
            "post": {
                "summary": "Get the travel data file of each of the given series of links in the specified time ranges.",
                "operationId": "getTravelDataFile",
                "requestBody": {
                    "description": "The date-time specifications and links(link_dirs) to be queried",
                    "required": true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/TravelFileRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The file containing the travel data array of the given series of links.",
                        "content": {
                            "text/plain; charset=utf-8": {
                                
                            }
                        }
                    }
                }
            }
        },
        "/date-bounds": {
            "get": {
                "summary": "Get the date time boundary of the data sets.",
                "operationId": "getDateBoundary",
                "responses": {
                    "200": {
                        "description": "The date time boundary of the data sets.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DateTimeBoundary"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TravelRequestBody": {
                "type": "object",
                "required": [
                    "start_time",
                    "end_time",
                    "link_dirs"
                ],
                "properties": {
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2018-09-30 19:55:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2018-09-30 19:55:00"
                    },
                    "link_dirs": {
                        "type": "array",
                        "items":{
                            "type": "string",
                            "example": "1242626341F"
                        }
                    }
                }
            },
            "TravelFileRequestBody": {
                "type": "object",
                "required": [
                    "start_date",
                    "end_date",
                    "start_time",
                    "end_time",
                    "days_of_week",
                    "include_holidays",
                    "link_dirs"
                ],
                "properties": {
                    "start_date": {
                        "type": "string",
                        "example": "2018-09-01"
                    },
                    "end_date": {
                        "type": "string",
                        "example": "2018-09-30"
                    },
                    "start_time": {
                        "type": "string",
                        "example":"00:00:00"
                    },
                    "end_time": {
                        "type": "string",
                        "example":"19:55:00"
                    },
                    "days_of_week": {
                        "type": "array",
                        "items": {
                            "type": "boolean"
                        },
                        "example": "[true, false, true, false, true, false, false]"
                    },
                    "include_holidays": {
                        "type": "boolean"
                    },
                    "link_dirs": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "1242626341F"
                        }
                    },
                    "file_type": {
                        "type": "string",
                        "example": "csv"
                    }
                }
            },
            "MultiLinkPath": {
                "type": "object",
                "required": [
                    "source",
                    "target",
                    "link_dirs",
                    "geometry"
                ],
                "properties": {
                    "source": {
                        "type": "integer"
                    },
                    "target": {
                        "type": "integer"
                    },
                    "link_dirs": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "29569307F"
                        }
                    },
                    "geometry": {
                        "type": "object",
                        "required": [
                            "type",
                            "coordinates"
                        ],
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "MultiLineString"
                            },
                            "coordinates": {
                                "type": "array",
                                "items": {
                                    "type": "array",
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "example": "[[[-79.38673, 43.64161], [-79.38658, 43.6417]], [[-79.38658, 43.6417], [-79.38632,43.64185]]]"
                            }
                        }
                    }
                }
            },
            "SingleLinkPath": {
                "type": "object",
                "required": [
                    "source",
                    "target",
                    "link_dirs",
                    "geometry"
                ],
                "properties": {
                    "source": {
                        "type": "integer"
                    },
                    "target": {
                        "type": "integer"
                    },
                    "link_dirs": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "29569307F"
                        }
                    },
                    "geometry": {
                        "type": "object",
                        "required": [
                            "type",
                            "coordinates"
                        ],
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "LineString"
                            },
                            "coordinates": {
                                "type": "array",
                                "items": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                },
                                "example": "[[-79.38658, 43.6417], [-79.38632, 43.64185]]"
                            }
                        }
                    }
                }
            },
            "Link": {
                "type": "object",
                "required": [
                    "link_dir",
                    "link_id",
                    "st_name",
                    "source",
                    "target",
                    "length",
                    "geometry"
                ],
                "properties": {
                    "link_dir": {
                        "type": "string",
                        "example": "1242626341F"
                    },
                    "link_id": {
                        "type": "integer"
                    },
                    "st_name": {
                        "type": "string"
                    },
                    "source": {
                        "type": "integer"
                    },
                    "target": {
                        "type": "integer"
                    },
                    "length": {
                        "type": "number"
                    },
                    "geometry": {
                        "type": "object",
                        "required": [
                            "type",
                            "coordinates"
                        ],
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "LineString"
                            },
                            "coordinates": {
                                "type": "array",
                                "items": {
                                    "type": "array",
                                    "items": {
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "example": "[ [ -79.4611, 43.72609 ], [ -79.46082, 43.72616 ] ]"
                            }
                        }
                    }
                }
            },
            "Node": {
                "type": "object",
                "required": [
                    "node_id",
                    "geometry"
                ],
                "properties": {
                    "node_id": {
                        "type": "integer"
                    },
                    "geometry": {
                        "type": "object",
                        "required": [
                            "type",
                            "coordinates"
                        ],
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "Point"
                            },
                            "coordinates": {
                                "type": "array",
                                "items": {
                                    "type": "number",
                                    "format": "double"
                                },
                                "example": "[ -79.38546, 43.64222 ]"
                            }
                        }
                    }
                }
            },
            "TravelData": {
                "type": "object",
                "required": [
                    "link_dir",
                    "tx",
                    "length",
                    "mean",
                    "stddev",
                    "confidence",
                    "pct_50"
                ],
                "properties": {
                    "link_dir": {
                        "type": "string",
                        "example": "1242626341F"
                    },
                    "tx": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2018-09-30 19:55:00"
                    },
                    "length": {
                        "type": "integer"
                    },
                    "mean": {
                        "type": "number",
                        "format": "double"
                    },
                    "stddev": {
                        "type": "number",
                        "format": "double"
                    },
                    "confidence": {
                        "type": "integer"
                    },
                    "pct_50": {
                        "type": "integer"
                    }
                }
            },
            "LinkNodesRequestBody": {
                "type":"object",
                "required":[
                    "node_ids"
                ],
                "properties": {
                    "node_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": "30310655"
                        }
                    }
                }
            },
            "DateTimeBoundary": {
                "type": "object",
                "required": [
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2018-09-30 19:55:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2018-09-30 19:55:00"
                    }
                }
            }
        }
    }
}